新建QT桌面工具工程 完成控制文件是否加密保存逻辑

+++ server/app.py
@@ -4,8 +4,9 @@ import random
 import shutil
 import datetime
 import base64
+import threading
 
-from Crypto.Cipher import AES
+from Crypto.Cipher import AES  # 安装依赖pycryptodome，不是Crypto
 from Crypto.Util import Counter
 import peewee
 from flask import Flask, jsonify, request, send_file
@@ -14,6 +15,11 @@ from itsdangerous import (TimedJSONWebSignatureSerializer as URLSafeSerializer, 
 from functools import wraps
 from flask_cors import CORS, cross_origin
 from playhouse.shortcuts import model_to_dict, dict_to_model
+import asyncio
+
+from watchdog.observers import Observer
+from watchdog.events import *
+from watchdog.utils.dirsnapshot import DirectorySnapshot, DirectorySnapshotDiff
 
 from DAO import *
 
@@ -21,17 +27,69 @@ app = Flask(__name__)
 app.config.from_object('config')
 # 允许跨域
 CORS(app, supports_credentials=True)
+SecreyKey = os.path.abspath(app.config['SECRET_KEY'])
+DesKey = os.path.abspath(app.config['DES_KEY'])
+ClouDiskPath = os.path.abspath(app.config['UPLOAD_FOLDER'])
+UpdateInterval = app.config['UPDATE_INTERVAL']
+UpdateMaxInterval = app.config['UPDATE_MAX_INTERVAL']
+Encrypt_File_Mode = app.config['ENCRYPT_FILE_MODE']
+
+
+class FileEventHandler(FileSystemEventHandler):
+    def __init__(self, cloud_disk_path):
+        FileSystemEventHandler.__init__(self)
+        self.cloud_disk_path = cloud_disk_path
+        self.timer_task = None
+        self.last_snapshot = DirectorySnapshot(self.cloud_disk_path)
+        self.cancel_count = 0
+
+    def on_any_event(self, event):
+        if self.timer_task:
+            self.timer_task.cancel()
+        self.timer_task = threading.Timer(UpdateInterval, self.check_snapshot)
+        self.timer_task.start()
+
+    def check_snapshot(self):
+        snapshot = DirectorySnapshot(self.cloud_disk_path)
+        diff = DirectorySnapshotDiff(self.last_snapshot, snapshot)
+        self.last_snapshot = snapshot
+        self.timer_task = None
+
+        print("files_created:", diff.files_created)
+        print("files_deleted:", diff.files_deleted)
+        print("files_modified:", diff.files_modified)
+        print("files_moved:", diff.files_moved)
+        print("dirs_modified:", diff.dirs_modified)
+        print("dirs_moved:", diff.dirs_moved)
+        print("dirs_deleted:", diff.dirs_deleted)
+        print("dirs_created:", diff.dirs_created)
+
+
+class DirMonitor(object):
+
+    def __init__(self, cloud_disk_path):
+        self.cloud_disk_path = cloud_disk_path
+        self.observer = Observer()
+
+    def start(self):
+        event_handler = FileEventHandler(self.cloud_disk_path)
+        print(event_handler)
+        self.observer.schedule(event_handler, self.cloud_disk_path, True)
+        self.observer.start()
+
+    def stop(self):
+        self.observer.stop()
 
 
 def verify_auth_token(token):
-    s = URLSafeSerializer(app.config['SECRET_KEY'])
+    s = URLSafeSerializer(SecreyKey)
     try:
         data = s.loads(token)
     except SignatureExpired:
         return None
     except BadSignature:
         return None
-    return 'key' in data and data['key'] == app.config['SECRET_KEY']
+    return 'key' in data and data['key'] == SecreyKey
 
 
 def test_authorization():
@@ -72,11 +130,19 @@ def generate_path(fid, pfolder, pchild):
     return actual_path
 
 
+def generate_foldername(folderid, foldername):
+    return hashlib.md5(("foldid" + folderid + "foldername" + foldername).encode('utf-8')).hexdigest()
+
+
+def generate_filename(folderid, filename):
+    return hashlib.md5(("foldid" + folderid + "filename" + filename).encode('utf-8')).hexdigest()
+
+
 def AES_Encrypt(data):
     vi = '0918273645abcdef'
     pad = lambda s: s + (16 - len(s) % 16) * chr(16 - len(s) % 16)
     data = pad(data)
-    cipher = AES.new(app.config['DES_KEY'].encode('utf8'), AES.MODE_CBC, vi.encode('utf8'))
+    cipher = AES.new(DesKey.encode('utf8'), AES.MODE_CBC, vi.encode('utf8'))
     encryptedbytes = cipher.encrypt(data.encode('utf8'))
     encodestrs = base64.urlsafe_b64encode(encryptedbytes)
     enctext = encodestrs.decode('utf8')
@@ -87,7 +153,7 @@ def AES_Decrypt(data):
     vi = '0918273645abcdef'
     data = data.encode('utf8')
     encodebytes = base64.urlsafe_b64decode(data)
-    cipher = AES.new(app.config['DES_KEY'].encode('utf8'), AES.MODE_CBC, vi.encode('utf8'))
+    cipher = AES.new(DesKey.encode('utf8'), AES.MODE_CBC, vi.encode('utf8'))
     text_decrypted = cipher.decrypt(encodebytes)
     unpad = lambda s: s[0:-s[-1]]
     text_decrypted = unpad(text_decrypted)
@@ -116,10 +182,26 @@ def get_decrypt_url(url):
     return srcdata
 
 
-def get_file_download_path(file):
+def get_encrypt_file_path(file):
+    actual_path = generate_filename(file.parentid, file.name)
+    target_file = os.path.join(os.path.expanduser(ClouDiskPath), actual_path)
+    return target_file
+
+
+def get_noencrypt_file_path(file):
     folderpath = Folder.get(Folder.id == file.parentid)
     actual_path = generate_path(file.parentid, folderpath.path, file.name)
-    target_file = os.path.join(os.path.expanduser(app.config['UPLOAD_FOLDER']), actual_path)
+    target_file = os.path.join(os.path.expanduser(ClouDiskPath), actual_path)
+    return target_file
+
+
+def get_download_file(file):
+    actual_path = generate_filename(file.parentid, file.name)
+    target_file = os.path.join(os.path.expanduser(ClouDiskPath), actual_path)
+    if not Encrypt_File_Mode:
+        folderpath = Folder.get(Folder.id == file.parentid)
+        actual_path = generate_path(file.parentid, folderpath.path, file.name)
+        target_file = os.path.join(os.path.expanduser(ClouDiskPath), actual_path)
     if os.path.exists(target_file):
         rv = send_file(target_file, as_attachment=True, attachment_filename=file.name)
         return rv
@@ -139,7 +221,6 @@ def base36_encode(number):
 
 def get_random_long_int():
     return random.SystemRandom().randint(1000000000, 9999999999)
-    # return _random.randint(1000000000, 9999999999)
 
 
 def get_random_short_int():
@@ -156,10 +237,10 @@ def login():
     # 验证用户名密码
     if req['email'] == app.config['EMAIL'] and req['password'] == app.config['PASSWORD']:
         # 生成 Token，有效期为一周
-        s = URLSafeSerializer(app.config['SECRET_KEY'], expires_in=7 * 24 * 3600)
+        s = URLSafeSerializer(SecreyKey, expires_in=7 * 24 * 3600)
 
         return jsonify(message='OK',
-                       token=s.dumps({'key': app.config['SECRET_KEY']}).decode('utf-8'))
+                       token=s.dumps({'key': SecreyKey}).decode('utf-8'))
     else:
         return jsonify(message='unauthorized'), 401
 
@@ -177,26 +258,35 @@ def folders():
         req = request.get_json()
         foldername = req['name']
         parentid = req['parentid']
+        if len(Folder.select().where((Folder.parentid == parentid) & (Folder.name == foldername))):
+            return jsonify(message='error'), 409
         folder = Folder.get_by_id(parentid)
-        # folderpath = foldername
-        # if parentid != 0:
-        #     folderpath = folder.path + '/' + foldername
-
         folderpath = generate_path(parentid, folder.path, foldername)
-        target_folder = os.path.join(os.path.expanduser(app.config['UPLOAD_FOLDER']), folderpath)
-        if os.path.exists(target_folder):
-            return jsonify(message='error'), 409
-        try:
-            os.mkdir(target_folder)
-            f = Folder.create(name=foldername, path=folderpath, parentid=parentid, isShared=False,
-                              isShareEncryped=False, sharePassword="", shareUrl="", sharePeriod=datetime.datetime.now())
-            f.save()
-            return jsonify(message='OK'), 201
-        except peewee.IntegrityError as e:
-            return jsonify(message='error'), 409
-        except Exception as e:
-            app.logger.exception(e)
-            return jsonify(message='error'), 409
+        if Encrypt_File_Mode:
+            try:
+                f = Folder.create(name=foldername, path=folderpath, parentid=parentid, isShared=False,
+                                  isShareEncryped=False, sharePassword="", shareUrl="",
+                                  sharePeriod=datetime.datetime.now())
+                f.save()
+                return jsonify(message='OK'), 201
+            except peewee.IntegrityError as e:
+                return jsonify(message='error'), 409
+        else:
+            target_folder = os.path.join(os.path.expanduser(ClouDiskPath), folderpath)
+            if os.path.exists(target_folder):
+                return jsonify(message='error'), 409
+            try:
+                os.mkdir(target_folder)
+                f = Folder.create(name=foldername, path=folderpath, parentid=parentid, isShared=False,
+                                  isShareEncryped=False, sharePassword="", shareUrl="",
+                                  sharePeriod=datetime.datetime.now())
+                f.save()
+                return jsonify(message='OK'), 201
+            except peewee.IntegrityError as e:
+                return jsonify(message='error'), 409
+            except Exception as e:
+                app.logger.exception(e)
+                return jsonify(message='error'), 409
     if request.method == 'GET':
         query = Folder.select()
         if query.exists():
@@ -224,7 +314,6 @@ def folder(folder_id):
                        files=[model_to_dict(file) for file in files])
     elif request.method == 'DELETE':
         folderpath = folder.path
-        target_file = os.path.join(os.path.expanduser(app.config['UPLOAD_FOLDER']), folderpath)
         try:
             # 递归删除数据库函数
             def recuesiondel(folder_id):
@@ -238,8 +327,10 @@ def folder(folder_id):
 
             # 递归删除数据库
             recuesiondel(folder_id)
-            # 删除文件夹
-            shutil.rmtree(target_file)
+            if not Encrypt_File_Mode:
+                target_file = os.path.join(os.path.expanduser(ClouDiskPath), folderpath)
+                # 删除文件夹
+                shutil.rmtree(target_file)
             # 数据库删除自身
             folder.delete_instance()
 
@@ -249,13 +340,15 @@ def folder(folder_id):
             app.logger.exception(e)
             return jsonify(message='error'), 409
     elif request.method == 'POST':
+        print(folder_id)
         file = request.files['file']
         if file:
-            # filepath = file.filename
-            # if folder_id != '0':
-            #     filepath = folder.path + '/' + file.filename
             filepath = generate_path(folder_id, folder.path, file.filename)
-            target_file = os.path.join(os.path.expanduser(app.config['UPLOAD_FOLDER']), filepath)
+            target_file = os.path.join(os.path.expanduser(ClouDiskPath), filepath)
+            if Encrypt_File_Mode:
+                filepath = generate_filename(folder_id, file.filename)
+                print(filepath)
+                target_file = os.path.join(os.path.expanduser(ClouDiskPath), filepath)
             if os.path.exists(target_file):
                 return jsonify(message='error'), 409
             try:
@@ -294,19 +387,18 @@ def files(file_id):
         file = File.get_by_id(file_id)
     except peewee.DoesNotExist:
         return jsonify(message='error'), 404
-    # actual_path = file.name
-    # if file.parentid != 0:
-    #     folderpath = Folder.select(Folder.path).where(Folder.id == file.parentid)
-    #     actual_path = folderpath + '/' + file.name
-    folderpath = Folder.get(Folder.id == file.parentid)
-    actual_path = generate_path(file.parentid, folderpath.path, file.name)
-    target_file = os.path.join(os.path.expanduser(app.config['UPLOAD_FOLDER']), actual_path)
+    actual_path = generate_filename(file.parentid, file.filename)
+    target_file = os.path.join(os.path.expanduser(ClouDiskPath), actual_path)
+    if not Encrypt_File_Mode:
+        folderpath = Folder.get(Folder.id == file.parentid)
+        actual_path = generate_path(file.parentid, folderpath.path, file.name)
+        target_file = os.path.join(os.path.expanduser(ClouDiskPath), actual_path)
     if request.method == 'GET':
         args = request.args
         if 'query' in args and args['query'] == 'info':
             return jsonify(message='OK', data=model_to_dict(file), actual_path=actual_path)
         if os.path.exists(target_file):
-            return send_file(target_file)
+            return send_file(target_file, as_attachment=True, attachment_filename=file.name)
         else:
             return jsonify(message='error'), 404
     elif request.method == 'DELETE':
@@ -345,7 +437,7 @@ def share(path):
     res = get_decrypt_url(path)
     args = request.args
     if res:
-        s = URLSafeSerializer(app.config['SECRET_KEY'], expires_in=24 * 3600)
+        s = URLSafeSerializer(SecreyKey, expires_in=24 * 3600)
 
         if res[0] == 'fi':
             try:
@@ -378,7 +470,7 @@ def share(path):
                             if data['path'] == path:
                                 payload['needPassword'] = False
                                 if 'download' in args and args['download'] == 'true':
-                                    rv = get_file_download_path(file)
+                                    rv = get_download_file(file)
                                     if rv:
                                         return rv
                                     else:
@@ -389,7 +481,7 @@ def share(path):
                             return jsonify(message='unauthorized', data=payload), 401
                 else:
                     if 'download' in args and args['download'] == 'true':
-                        rv = get_file_download_path(file)
+                        rv = get_download_file(file)
                         if rv:
                             return rv
                         else:
@@ -407,4 +499,6 @@ def share(path):
 
 
 if __name__ == '__main__':
-    app.run(debug=True)
+    # monitor = DirMonitor(ClouDiskPath)
+    # monitor.start()
+    app.run(debug=True, use_reloader=False)
